cleanup

+++ 2024/day13/part1.cpp
@@ -70,8 +70,8 @@ void parse(string const& line, int & x, int & y) {
 
 	regex_search(line, match, re);
 
-	x = stof(match[1].str());
-	y = stof(match[2].str());
+	x = stol(match[1].str());
+	y = stol(match[2].str());
 
 }
